refactor: set good namespaces for some response models
