+++ src/Shared/DTO/Restricted/AuthResponseModel.cs
@@ -1,4 +1,4 @@
-namespace Iso.API.DTO;
+namespace Iso.Shared.DTO.Restricted;
 
 public record AuthResponseModel(
     string token);
